requirements.txt


allure-pytest==2.8.33
allure-python-commons==2.8.33
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
boto3==1.17.12
botocore==1.20.12
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.3
Django==3.1.7
django-basic-stats==0.2.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
gcloud==0.18.3
gevent==21.1.2
geventhttpclient==1.4.5
google-auth==1.26.1
googleapis-common-protos==1.53.0
greenlet==1.0.0
httplib2==0.19.0
idna==2.10
iniconfig==1.1.1
invokust==0.76
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
kubernetes==12.0.1
locust==1.4.3
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0
npm==0.1.1
numpy==1.20.1
oauth2client==4.1.3
oauthlib==3.1.0
optional-django==0.1.0
packaging==20.9
pluggy==0.13.1
protobuf==3.15.6
psutil==5.8.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
stats==0.1.2a0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
websocket-client==0.57.0
Werkzeug==1.0.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0